Specify QGIS argument types

Specify QGIS argument types

Usage

qgis_argument_spec(
  algorithm = NA_character_,
  name = NA_character_,
  description = NA_character_,
  qgis_type = NA_character_,
  available_values = character(0),
  acceptable_values = character(0)
)

Arguments

algorithm

A qualified algorithm name (e.g., "native:buffer") or a path to a QGIS model file.

name, description, qgis_type, available_values, acceptable_values

Column values of arguments denoting the argument name, description, and acceptable values.

Value

A list() with an element for each input argument.

Examples

# These became internal functions!
qgisprocess:::qgis_argument_spec()
#> $algorithm
#> [1] NA
#> 
#> $name
#> [1] NA
#> 
#> $description
#> [1] NA
#> 
#> $qgis_type
#> [1] NA
#> 
#> $available_values
#> character(0)
#> 
#> $acceptable_values
#> character(0)
#> 
if (has_qgis()) qgisprocess:::qgis_argument_spec_by_name("native:filedownloader", "URL")
#> $algorithm
#> [1] "native:filedownloader"
#> 
#> $name
#> [1] "URL"
#> 
#> $description
#> [1] "URL"
#> 
#> $qgis_type
#> [1] "string"
#> 
#> $default_value
#> NULL
#> 
#> $available_values
#> NULL
#> 
#> $acceptable_values
#> [1] "String value"                                                                                    
#> [2] "field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field"                    
#> [3] "expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression"
#>