Configure qgisprocess

Run qgis_configure() to bring the package configuration in line with QGIS and to save this configuration to a persistent cache. See the Details section for more information about setting the path of the 'qgis_process' command line tool.

Usage

qgis_configure(quiet = FALSE, use_cached_data = FALSE)

Arguments

quiet

Use FALSE to display more information, possibly useful for debugging.

use_cached_data

Use the cached algorithm list and path found when configuring qgisprocess during the last session. This saves some time loading the package.

Value

The result of processx::run().

Details

The qgisprocess package is a wapper around the 'qgis_process' command line tool distributed with QGIS (>=3.14). Several functions use heuristics to detect the location of the 'qgis_process' executable.

When loading the package, the configuration is automatically read from the cache with qgis_configure(use_cached_data = TRUE, quiet = TRUE) in order to save time. Run qgis_configure(use_cached_data = TRUE) manually to get more details.

Use qgis_algorithms(), qgis_providers(), qgis_plugins(), qgis_using_json_output(), qgis_path() and qgis_version() to inspect cache contents.

If the configuration fails or you have more than one QGIS installation, you can set options(qgisprocess.path = "path/to/qgis_process") or the R_QGISPROCESS_PATH environment variable (useful on CI). On Linux the 'qgis_process' executable is generally available on the user's PATH, on MacOS the executable is within the QGIS*.app/Contents/MacOS/bin folder, and on Windows the executable is named qgis_process-qgis.bat or qgis_process-qgis-dev.bat and is located in Program Files/QGIS*/bin or OSGeo4W(64)/bin.

See also

Examples

if (has_qgis()) qgis_path()
#> [1] "qgis_process"
if (has_qgis()) qgis_version()
#> [1] "3.30.2-'s-Hertogenbosch"
if (has_qgis()) qgis_algorithms()
#> # A tibble: 349 × 24
#>    provider provider_title algorithm                algorithm_id algorithm_title
#>    <chr>    <chr>          <chr>                    <chr>        <chr>          
#>  1 3d       QGIS (3D)      3d:tessellate            tessellate   Tessellate     
#>  2 gdal     GDAL           gdal:aspect              aspect       Aspect         
#>  3 gdal     GDAL           gdal:assignprojection    assignproje… Assign project…
#>  4 gdal     GDAL           gdal:buffervectors       buffervecto… Buffer vectors 
#>  5 gdal     GDAL           gdal:buildvirtualraster  buildvirtua… Build virtual …
#>  6 gdal     GDAL           gdal:buildvirtualvector  buildvirtua… Build virtual …
#>  7 gdal     GDAL           gdal:cliprasterbyextent  cliprasterb… Clip raster by…
#>  8 gdal     GDAL           gdal:cliprasterbymaskla… cliprasterb… Clip raster by…
#>  9 gdal     GDAL           gdal:clipvectorbyextent  clipvectorb… Clip vector by…
#> 10 gdal     GDAL           gdal:clipvectorbypolygon clipvectorb… Clip vector by…
#> # ℹ 339 more rows
#> # ℹ 19 more variables: provider_can_be_activated <lgl>,
#> #   provider_is_active <lgl>, provider_long_name <chr>, provider_version <chr>,
#> #   provider_warning <chr>, can_cancel <lgl>, deprecated <lgl>, group <chr>,
#> #   has_known_issues <lgl>, help_url <chr>, requires_matching_crs <lgl>,
#> #   short_description <chr>, tags <list>, default_raster_file_extension <chr>,
#> #   default_vector_file_extension <chr>, …
if (has_qgis()) qgis_providers()
#> # A tibble: 4 × 3
#>   provider provider_title    algorithm_count
#>   <chr>    <chr>                       <int>
#> 1 gdal     GDAL                           56
#> 2 qgis     QGIS                           50
#> 3 3d       QGIS (3D)                       1
#> 4 native   QGIS (native c++)             242
if (has_qgis()) qgis_plugins()
#> # A tibble: 3 × 2
#>   name          enabled
#>   <chr>         <lgl>  
#> 1 grassprovider FALSE  
#> 2 otbprovider   FALSE  
#> 3 processing    TRUE   
qgis_configure(use_cached_data = TRUE)
#> Checking configuration in cache file (~/.cache/R-qgisprocess/cache-0.0.0.9000.rds)
#> Checking cached QGIS version with version reported by 'qgis_process' ...
#> QGIS versions match! (3.30.2-'s-Hertogenbosch)
#> Checking cached QGIS plugins (and state) with those reported by 'qgis_process' ...
#> QGIS plugins match! (1 processing provider plugin(s) enabled)
#> 
#> >>> Run `qgis_enable_plugins()` to enable 2 disabled plugin(s) and
#>     access their algorithms: grassprovider, otbprovider
#> 
#> Restoring configuration from '~/.cache/R-qgisprocess/cache-0.0.0.9000.rds'
#> QGIS version: 3.30.2-'s-Hertogenbosch
#> Using 'qgis_process' in the system PATH.
#> >>> If you need another installed QGIS instance, run `qgis_configure()`;
#>     see `?qgis_configure` if you need to preset the path of 'qgis_process'.
#> Using JSON for output serialization.
#> Using JSON for input serialization.
#> 1 out of 3 available processing provider plugins are enabled.
#> Having access to 349 algorithms from 4 QGIS processing providers.
#> Use qgis_algorithms(), qgis_providers(), qgis_plugins(), qgis_using_json_output(),
#> qgis_path() and qgis_version() to inspect the cache environment.
qgis_configure()
#> getOption('qgisprocess.path') was not found.
#> Sys.getenv('R_QGISPROCESS_PATH') was not found.
#> Trying 'qgis_process' on PATH...
#> Success!
#> Now using 'qgis_process' in the system PATH.
#> >>> If you need another installed QGIS instance, run `qgis_configure()`;
#>     see `?qgis_configure` if you need to preset the path of 'qgis_process'.
#> 
#> QGIS version is now set to: 3.30.2-'s-Hertogenbosch
#> Using JSON for output serialization.
#> Using JSON for input serialization.
#> 1 out of 3 available processing provider plugins are enabled.
#> You now have access to 349 algorithms from 4 QGIS processing providers.
#> 
#> >>> Run `qgis_enable_plugins()` to enable 2 disabled plugin(s) and
#>     access their algorithms: grassprovider, otbprovider
#> 
#> Saving configuration to '~/.cache/R-qgisprocess/cache-0.0.0.9000.rds'
#> Use qgis_algorithms(), qgis_providers(), qgis_plugins(), qgis_using_json_output(),
#> qgis_path() and qgis_version() to inspect the cache environment.