Create functions from QGIS algorithms

As opposed to qgis_run_algorithm(), qgis_function() creates a callable function based on the argument metadata provided by qgis_arguments(). Unlike qgis_run_algorithm(), qgis_function() sets the default value of .quiet to TRUE to make the function more usable within other R code. Similarly, qgis_pipe() wraps qgis_run_algorithm(), passing its first argument to the first input to algorithm.

qgis_function(algorithm, ...)

qgis_pipe(.data, algorithm, ..., .quiet = TRUE)

Arguments

algorithm

A qualified algorithm name (e.g., "native:filedownloader") or a path to a QGIS model file.

...

Default values to set when using qgis_function(). These values are evaluated once and immediately, so you shouldn't call qgis_tmp_file() here.

.data

Passed to the first input of algorithm.

.quiet

Use TRUE to suppress output from processing algorithms.

Examples

if (has_qgis()) {
  qgis_buffer <- qgis_function("native:buffer")
  qgis_buffer(
    system.file(
      "longlake/longlake_depth.gpkg",
      package = "qgisprocess"
    ),
    DISTANCE = 10
  )
}
#> Argument `SEGMENTS` is unspecified (using QGIS default value).
#> Using `END_CAP_STYLE = "Round"`
#> Using `JOIN_STYLE = "Round"`
#> Argument `MITER_LIMIT` is unspecified (using QGIS default value).
#> Argument `DISSOLVE` is unspecified (using QGIS default value).
#> Using `OUTPUT = qgis_tmp_vector()`
#> <Result of `qgis_run_algorithm("native:buffer", ...)`>
#> List of 1
#>  $ OUTPUT: 'qgis_outputVector' chr "/tmp/RtmpGJBJsw/file2b9b4bf99094/file2b9b7eb75d11.gpkg"

if (has_qgis()) {
  qgis_pipe(
    system.file(
      "longlake/longlake_depth.gpkg",
      package = "qgisprocess"
    ),
    "native:buffer",
    DISTANCE = 10
  )
}
#> Argument `SEGMENTS` is unspecified (using QGIS default value).
#> Using `END_CAP_STYLE = "Round"`
#> Using `JOIN_STYLE = "Round"`
#> Argument `MITER_LIMIT` is unspecified (using QGIS default value).
#> Argument `DISSOLVE` is unspecified (using QGIS default value).
#> Using `OUTPUT = qgis_tmp_vector()`
#> <Result of `qgis_run_algorithm("native:buffer", ...)`>
#> List of 1
#>  $ OUTPUT: 'qgis_outputVector' chr "/tmp/RtmpGJBJsw/file2b9b4bf99094/file2b9b3f031e3f.gpkg"