List, enable or disable QGIS plugins

qgis_plugins() lists the available plugins that implement Processing providers. qgis_enable_plugins() enables plugins while qgis_disable_plugins() does the reverse.

qgis_plugins(which = "all", query = FALSE, quiet = TRUE, ...)

qgis_enable_plugins(names = NULL, quiet = FALSE)

qgis_disable_plugins(names = NULL, quiet = FALSE)

Arguments

which

String defining which plugins to select, based on their status in QGIS (enabled or disabled). Must be one of: "all", "enabled", "disabled".

query

Use TRUE to refresh the cached value.

quiet

Use FALSE to display more information about the command, possibly useful for debugging.

...

Only used by other functions calling this function.

names

Optional character vector of plugin names.

Value

A tibble of plugins and their status.

Details

The cache is immediately updated upon enabling or disabling plugins from R.

If you installed, removed, enabled or disabled plugins in the QGIS GUI, then run qgis_configure() to make those changes available in R.

If names is not provided to qgis_enable_plugins(), it is assumed that all disabled plugins are to be enabled. If names is not provided to qgis_disable_plugins(), it is assumed that all enabled plugins are to be disabled. Note that the 'processing' plugin is ignored, because it is always available to 'qgis_process' (not QGIS though).

Note

Only plugins that implement Processing providers are supported. Installing or removing plugins is not supported.