Configure and run 'qgis_process'

The qgisprocess package is a wapper around the 'qgis_process' command line tool distributed with QGIS (>=3.14). These functions use heuristics to detect the location of the 'qgis_process' executable. If the configuration fails or you have more than one QGIS installation, you can set options(qgisprocess.path = "path/to/qgis_process") or the R_QGISPROCESS_PATH environment variable (useful on CI). On Linux the 'qgis_process' executable is generally available on the user's PATH, on MacOS the executable is within the QGIS*.app/Contents/MacOS/bin folder, and on Windows the executable is named qgis_process-qgis.bat or qgis_process-qgis-dev.bat and is located in Program Files/QGIS*/bin or OSGeo4W(64)/bin.

qgis_run(args = character(), ..., env = qgis_env(), path = qgis_path())

has_qgis()

assert_qgis(action = abort)

qgis_configure(quiet = FALSE)

qgis_unconfigure()

qgis_version(query = FALSE, quiet = TRUE)

qgis_path(query = FALSE, quiet = TRUE)

qgis_query_path(quiet = FALSE)

qgis_env()

qgis_query_version(quiet = FALSE)

qgis_query_algorithms(quiet = FALSE)

Arguments

args
Command-line arguments

...
Passed to processx::run().

env
A list() of environment variables. Defaults to qgis_env().

path
A path to the 'qgis_process' executable. Defaults to qgis_path().

action
An action to take if the 'qgis_process' executable could not be found.

quiet
Use FALSE to display more information about the command, possibly useful for debugging.

query
Use TRUE to refresh the cached value.

Value

The result of processx::run().

Examples

if (has_qgis()) qgis_path()

#> [1] "/Applications/QGIS.app/Contents/MacOS/bin/qgis_process"

if (has_qgis()) qgis_version()

#> [1] "3.20.1-Odense"

if (has_qgis()) qgis_algorithms()

#> # A tibble: 989 × 5
#>    provider provider_title algorithm                  algorithm_id algorithm_title
#>    <chr>    <chr>          <chr>                      <chr>        <chr>          
#>  1 3d       QGIS (3D)      3d:tessellate              tessellate   Tessellate     
#>  2 gdal     GDAL           gdal:aspect                aspect       Aspect         
#>  3 gdal     GDAL           gdal:assignprojection      assignproje… Assign project…
#>  4 gdal     GDAL           gdal:buffervectors         buffervecto… Buffer vectors 
#>  5 gdal     GDAL           gdal:buildvirtualraster    buildvirtua… Build virtual …
#>  6 gdal     GDAL           gdal:buildvirtualvector    buildvirtua… Build virtual …
#>  7 gdal     GDAL           gdal:cliprasterbyextent    cliprasterb… Clip raster by…
#>  8 gdal     GDAL           gdal:cliprasterbymasklayer cliprasterb… Clip raster by…
#>  9 gdal     GDAL           gdal:clipvectorbyextent    clipvectorb… Clip vector by…
#> 10 gdal     GDAL           gdal:clipvectorbypolygon   clipvectorb… Clip vector by…
#> # … with 979 more rows

qgis_configure()

#> getOption('qgisprocess.path') was not found.

#> Sys.getenv('R_QGISPROCESS_PATH') was not found.

#> Trying 'qgis_process' on PATH

#> Error in rethrow_call(c_processx_exec, command, c(command, args), pty, : cannot start processx process 'qgis_process' (system error 2, No such file or directory) @unix/processx.c:610 (processx_exec)

#> Found 1 QGIS installation containing 'qgis_process':
#>  /Applications/QGIS.app/Contents/MacOS/bin/qgis_process

#> Trying command '/Applications/QGIS.app/Contents/MacOS/bin/qgis_process'

#> Success!

#> QGIS version: 3.20.1-Odense

#> Metadata of 989 algorithms queried and stored in cache.
#> Run `qgis_algorithms()` to see them.