Type coercion for arguments to QGIS processing algorithms

Calls to qgis_run_algorithm() can and should contain R objects that need to be serialized before they are passed to the command line. In some cases (e.g., sf objects), temporary files need to be written and cleaned up. The as_qgis_argument() and qgis_clean_argument() S3 generics provide hooks for argument values to be serialized correctly.

qgis_sanitize_arguments(
  algorithm,
  ...,
  .algorithm_arguments = qgis_arguments(algorithm),
  .use_json_input = FALSE
)

qgis_clean_arguments(arguments)

as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_default_value()

is_qgis_default_value(x)

# S3 method for default
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for list
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for qgis_default_value
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for `NULL`
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for character
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for logical
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

# S3 method for numeric
as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_clean_argument(value)

# S3 method for default
qgis_clean_argument(value)

# S3 method for qgis_tempfile_arg
qgis_clean_argument(value)

qgis_argument_spec_by_name(
  algorithm,
  name,
  .algorithm_arguments = qgis_arguments(algorithm)
)

Arguments

algorithm

A qualified algorithm name (e.g., "native:filedownloader") or a path to a QGIS model file.

...

Named key-value pairs as arguments for each algorithm. Features of rlang::list2() are supported. These arguments are converted to strings using as_qgis_argument().

.algorithm_arguments

The result of qgis_arguments()

.use_json_input, use_json_input

TRUE if the outputs will be serialized as JSON instead of serialized as a command-line argument.

arguments

The result of qgis_sanitize_arguments().

x

An object passed to a QGIS processing algorithm

spec

A list() with values for algorithm, name, description, and qgis_type. See qgis_argument_spec() to create a blank spec for testing.

value

The result of as_qgis_argument() after the QGIS processing algorithm has been run.

name

The name of the input argument