Get detailed information about one algorithm

Get detailed information about one algorithm

Usage

qgis_show_help(algorithm)

qgis_get_description(algorithm)

qgis_get_argument_specs(algorithm)

qgis_get_output_specs(algorithm)

Arguments

algorithm

A qualified algorithm name (e.g., "native:buffer").

See also

Examples

if (has_qgis()) qgis_show_help("native:filedownloader")
#> Download file (native:filedownloader)
#> 
#> ----------------
#> Description
#> ----------------
#> This algorithm downloads a URL on the file system.
#> 
#> ----------------
#> Arguments
#> ----------------
#> 
#> URL: URL
#> 	Argument type:	string
#> 	Acceptable values:
#> 		- String value
#> 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field
#> 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
#> METHOD: Method
#> 	Default value:	0
#> 	The HTTP method to use for the request
#> 	Argument type:	enum
#> 	Available values:
#> 		- 0: GET
#> 		- 1: POST
#> 	Acceptable values:
#> 		- Number of selected option, e.g. '1'
#> 		- Comma separated list of options, e.g. '1,3'
#> DATA: Data (optional)
#> 	The data to add in the body if the request is a POST
#> 	Argument type:	string
#> 	Acceptable values:
#> 		- String value
#> 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field
#> 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
#> OUTPUT: File destination (optional)
#> 	Argument type:	fileDestination
#> 	Acceptable values:
#> 		- Path for new file
#> 
#> ----------------
#> Outputs
#> ----------------
#> 
#> OUTPUT: <outputFile>
#> 	File destination
#> 
#> 
#> 
if (has_qgis()) qgis_get_description("native:filedownloader")
#>                                native:filedownloader 
#> "This algorithm downloads a URL on the file system." 
if (has_qgis()) qgis_get_argument_specs("native:filedownloader")
#> # A tibble: 4 × 6
#>   name   description  qgis_type default_value available_values acceptable_values
#>   <chr>  <chr>        <chr>     <list>        <list>           <list>           
#> 1 URL    URL          string    <NULL>        <NULL>           <chr [3]>        
#> 2 METHOD Method       enum      <int [1]>     <chr [2]>        <chr [2]>        
#> 3 DATA   Data         string    <NULL>        <NULL>           <chr [3]>        
#> 4 OUTPUT File destin… fileDest… <NULL>        <NULL>           <chr [1]>