Manage temporary files

These functions create temporary files that can be used in calls to qgis_run_algorithm() or elsewhere. These files are created in a special temporary directory (qgis_tmp_base()) that should be periodically cleaned up using qgis_tmp_clean(). You can set your preferred vector and/or raster file extension using options(qgisprocess.tmp_vector_ext = "...") and/or options(qgisprocess.tmp_raster_ext = "..."), respectively.

qgis_tmp_file(ext)

is_qgis_tmp_file(x)

qgis_tmp_folder()

qgis_tmp_vector(ext = getOption("qgisprocess.tmp_vector_ext", ".gpkg"))

qgis_tmp_raster(ext = getOption("qgisprocess.tmp_raster_ext", ".tif"))

qgis_tmp_base()

qgis_tmp_clean()

Arguments

ext

The file extension to be used.

x

A character vector of file names.

Value

A character vector indicating the location of a (not yet created) temporary file.

Examples

qgis_tmp_base()
#> [1] "/tmp/RtmpcQjGAC/file626220f383d5"
qgis_tmp_file(".csv")
#> [1] "/tmp/RtmpcQjGAC/file626220f383d5/file62622e958ec6.csv"
qgis_tmp_vector()
#> [1] "/tmp/RtmpcQjGAC/file626220f383d5/file626264b69c6d.gpkg"
qgis_tmp_raster()
#> [1] "/tmp/RtmpcQjGAC/file626220f383d5/file626235ef362e.tif"