qgisprocess: Use 'QGIS' Processing Algorithms

Provides seamless access to the 'QGIS' (https://qgis.org/en/site/) processing toolbox using the standalone 'qgis_process' command-line utility. Both native and third-party (plugin) processing providers are supported. Beside referring data sources from file, also common objects from 'sf', 'terra' and 'stars' are supported. The native processing algorithms are documented by QGIS.org (2023) https://docs.qgis.org/latest/en/docs/user_manual/processing_algs/.

Note

A structured overview of the available functions can be consulted at https://r-spatial.github.io/qgisprocess/reference/index.html.

See also

Author

Maintainer: Floris Vanderhaeghe floris.vanderhaeghe@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Authors:

Dewey Dunnington dewey@fishandwhistle.net (ORCID) (Voltron Data)

Jan Caha jan.caha@outlook.com (ORCID)

Jannes Muenchow malnamalja@gmx.de (ORCID)

Other contributors:

Antony Barja antony.barja8@gmail.com (ORCID) [contributor]

Robin Lovelace rob00x@gmail.com (ORCID) [contributor]

Jakub Nowosad nowosad.jakub@gmail.com (ORCID) [contributor]

Research Institute for Nature and Forest (INBO) (https://www.inbo.be/en/) [copyright holder, funder]