Convert a qgis_result object or one of its elements to a raster object

Convert a qgis_result object or one of its elements to a raster object

Usage

qgis_as_raster(output, ...)

qgis_as_brick(output, ...)

# S3 method for qgis_outputRaster
qgis_as_raster(output, ...)

# S3 method for qgis_outputRaster
qgis_as_brick(output, ...)

# S3 method for qgis_outputLayer
qgis_as_raster(output, ...)

# S3 method for qgis_outputLayer
qgis_as_brick(output, ...)

# S3 method for qgis_result
qgis_as_raster(output, ...)

# S3 method for qgis_result
qgis_as_brick(output, ...)

Arguments

output

A result from qgis_run_algorithm() or one of the qgis_extract_output() functions.

...

Arguments passed to raster::raster() or raster::brick().

Note

Just use qgis_as_raster() in R scripts, it will use the correct method.

See also